key: fix KeyVault name
Prefixing KeyVault with cluster ID is risky because of naming
constraints. KeyVault name must not start with a number. The name has to
be globally unique.

I decided to prefix it (and eventually all other resources) with gs- to
guarantee we always start with a letter and avoid such situations in the
future.