Introduce state machine for instance resource
Instance resource has been working with persisted state for a long time
already. The state is kept in CR's status condition and updated accordingly as
instances upgrade proceeds. This has been necessary because in Azure the VMSS
instances upgrade process requires multiple steps for each VM and it's not
automated in similar way as the ASG in AWS.

Now that there needs to be more steps added into this process in order to make
upgrade more graceful for workload, maintenance and reasoning of this resource
became difficult during development process.

Hence this change adds a concept of resource internal state machine that drives
the resource behavior when reconciling a cluster upgrade.

Motivation behind this design is to make multi-step nature of instances upgrade
a first class citizen while maintaining cohesion for azure-operator overall.
This also makes certain level of Azure API call optimization possible when it
otherwise could not be so straightforward.