Improve Ingress LB PIP name (#604)
Original name for ingress LB public IP was just a "dummy-pip". Now that k8s
provider integration has actually support for re-using that on services with
type `LoadBalancer`, it makes sense to use more descriptive name for that.

In order to manage the migration of this name, store it in CR status. Older
clusters, which are migrated into this version, should still use the old name
as that cannot be changed, but newer ones must use the more descriptive one.

When the name of the public IP is in the status, `deployment` resource will then
pick up the name of already allocated public IP and assign it as an annotation
to ingress LB.