Only roll nodes when they aren't in sync with provider operator (#748)
It is difficult to detect different change scenarious with Azure APIs /
infrastructure and that has invoked full upgrade process e.g. when creating
cluster or scaling despite of many checks that try to test for such situation.

On principal rolling of node should not be needed other than when clusters are
upgrading or when certificates need renewal. This change introduces first of
these so upgrade process only executes through when there are nodes with old
azure-operator version labels. Similar case for expiring certificates should be
added later if needed.