Simplify master nodes' upgrade process
When upgrading master nodes, it is not needed to first update the VM
instance (this triggers reboot), then drain the node and eventually
re-image (this reboots the node again) but it should be sufficient to
just directly proceed with reimaging given that all nodes are `Ready` in
k8s api and VMSS instances are in `Successful` state.

This is expected to make master nodes upgrade a lot faster while
maintaining same disruption level as before. Because the initial VM
update rebooted the node already, the following draining doesn't have
any positive effect and hence this whole process can be removed.

Towards https://github.com/giantswarm/giantswarm/issues/13847