Roll nodes on major changes during release upgrade
When certain critical / major components such as kubernetes or flatcar
version changes during release upgrade, roll the node. Typical example
of this requirement is to apply security fix with release upgrade. With
current giant swarm approach, the kubernetes version is "baked" into
ignition template so VMs must be recycled in order to apply the patch.

This change adds release version label to cluster nodes and it's
currently not updated if release changes don't require rolling nodes.
This is preferable so that when there is next upgrade[s] in the future,
the change detection knows the components used to build the given node.
In case of release absence, nodes are always rolled.