Implement releasing of allocated subnet on deletion
When deleting node pool, the subnet allocation must be released in order
to make it available for later node pools.

This change adds an interface for `Releaser` and NOP implementation for
AzureConfig (because cluster resource allocations are released on
deletion automatically) and default implementation for
`AzureMachinePool` which removes the allocated subnet from
`AzureCluster` subnets.