Release allocated node pool subnet on deletion
* Refactor ipam.Check() to return existing subnet (#1181)

As a preparation for fixing subnet releasing on deletion of node pool,
refactor `ipam.Check()` to return allocated subnet if it exists.

* Implement releasing of allocated subnet on deletion (#1183)

When deleting node pool, the subnet allocation must be released in order
to make it available for later node pools.

This change adds an interface for `Releaser` and NOP implementation for
AzureConfig (because cluster resource allocations are released on
deletion automatically) and default implementation for
`AzureMachinePool` which removes the allocated subnet from
`AzureCluster` subnets.

Co-authored-by: Jose Armesto <github@armesto.net>