Implement migration for worker network fw rules (#1193)
Due to change prevention in network security group rules (to guard
customers' custom rules) the firewall rules don't get updated on cluster
upgrade and therefore the source and destination CIDR for some
rules is wrong.

This change implements a migration path to replace built-in worker
subnet CIDR with VNET one.